Prevent payload appearing in worker label
The `job.name` will return the underlying job class, along with the method
signature and payload. If the payload happens to contain PII it will appear in
the /metrics endpoint.

Instead, for ActiveJob we need to loiok at the `payload_object.job_data` in
order to get the job class by itself.

If ActiveJob is not present we'll take the `payload_object.class` for now
(which is always `Delayed::PerformableMethod`).